Fix model crash due to out-of-bound array access in physics data
When `base_idx` cannot be found by the `do` loop labeled `base_idx_loop`
above, its value is set to `size(base_var_names) + 1`. Then, a few lines
after, it is used again to access `base_var_names(base_idx)`, which
constitutes an out-of-bound array access. CAM-SIMA will crash due to this.

The solution is to `return` the subroutine early, and do not proceed any
further.